requirements.txt
flask==2.3.3
google-generativeai==0.3.2
chromadb==0.4.15
sentence-transformers>=2.2.2
PyPDF2==3.0.1
python-docx==0.8.11
langchain==0.0.350
langchain-google-genai==0.0.8
python-dotenv==1.0.0
gunicorn==21.2.0
spacy>=3.6.0
networkx>=3.1
numpy>=1.24.0
scikit-learn>=1.3.0
jieba>=0.42.1
opencc-python-reimplemented>=0.1.7
